client/requirements.txt
tensorflow-federated==0.87.0
tensorflow==2.14.1
tensorflow_privacy==0.9.0
paho-mqtt==2.1.0
pandas==2.2.3